requirements.txt
rich
aiosqlite
discord.py
python-dotenv